Add files via upload
This is an implementation of a Digital Multimeter on an STM32F407 Discovery board, which uses a Cortex M4 Processor. The Code is split into different sections and libraries, with physical pushbuttons being used as inputs, and a 16x2 LED as the output. ADC is used for the actual processes. The DMM can measure voltage, current, and resistance, all via the ADC using physical circuitry. All the coding has been done in C++.